Adds Window Layout features and control functions.
This commit adds two new lua functions `loadWindowLayout()` and
`saveWindowLayout()` as well as modifies `openUserWindow( name )` with
an optional second boolean argument to allow scripters manual control
over the layout features if they need it.
When a window is opened the layout information will automatically be
reloaded and applied if available, unless called like so:
`openUserWindow( 'my window', false )`

A CheckBox has been added to the Quit prompt which allows a user to
indicate if they would like to update the window layout before closing
the profile.

This code also changes/completes code in `Host.cpp` which is used to
mark a Host as closing or check its closing state.
The code is used in the `TConsole::closeEvent` and in
`TDockWidget::closeEvent` to aid in closing and removing the consoles
from Mudlet when a Host profile tab is closed.