Revise: correct clang-format disable/enable comments so they work [ci skip]
The tables containing the encoding of Extended ASCII character codes to
the representation as Unicode code-points for "other" codecs that
ISO-8859-1 are laid out carefully so that individual entries can be
identified as best they can.  However the current (any?) clang-format
configuration will destroy the layout so I had deliberately marked them
as not be touched by that tool. Unfortunately I got the comments to do this
slightly wrong so they did not work...!

As there is no functional code change here I have marked this commit to NOT
induce a Travis C.I. run.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>