BugFix: clean-up SIGNAL usage of and properly reuse QMediaPlayer instances
Although a newly created QMediaPlayer is created on demand when needed to
play a sound file (and all the existing instances are in use) the SIGNAL
that it creates when it has finished playing was tied to a lambda function
that creates a "sysSoundFinished" TEvent that is sent to the Host/profile
that initiated the sound file playback. However as the QMediaPlayer is then
available to be used for playback by ANY profile that next wants a sound
played the connection to that SIGNAL needs to be "rewired" to the
Host/profile that next uses it.  This commit restructures the code so this
happens every time mudlet::playSound(...) is used - the call to
disconnect(...) preceeding repositioned connect(...) will remove any
connection (but does not have a problem if there is not one).

I do not know enough about lambda functions to be certain but presumably
the memory allocated for the one in the connect(...) code will also get
release/freed when the connection is destroyed?

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>