Revision: changes suggested during review.
Changed to use a C++11 ranged "for" instead of Qt macro "foreach" in
TBuffer constructor.

Changed format of run-time error for an invalid "encoding" string supplied to
to the Lua setServerEncoding("encoding") function.

Reordered encodings list so that UTF-8 follows immediately after ASCII
rather than being on the end of the list.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>