BugFix: remove corruption in UTF-8 mode caused by off-by-one error
If the last bytes of a packet are multi-byte UTF-8 sequence it is necessary
to store them if there is not enough bytes in the packet to complete the
processing of them in TBuffer::translateToPlainText(...) unfortunately I
had made an error in assessing whether there WAS enough bytes so that the
code to store the incomplete bytes was being triggered when the last
character in the packet exactly ended at the end of the packet, i.e. it DID
fit in!

Also:
* Added a couple of QDebug()s - conditional on DEBUG_UTF8_PROCESSING that:
  report that an incomplete sequence has been detected and how many bytes
  are being stored; and report when they are being prepended onto the data
  in the next call to the method concerned.
* Revised the code that prepended those bytes to slightly optimise the
  tests that detect it is needed.
* Whilst using Mudlet replays to debug this issue I noticed that the
  replay starting and ending messages did not have the classification tag
  of the form "[ XXXX ] -" so did not display in any special way, I revised
  them to become an "[ INFO ]" message for replay starting and "[  OK  ]"
  at the end...!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>