Revise: reorder QGridLayout items in profile preferences form
This should reduce the signal to noise ratio when squashed into the other
commits for the PR...!

Also it restores the currently NOW empty "groupBox_Debug" on the last tab
of   the form. Although it has been emptied as a result of other commits in
the series of commits it is useful to have somewhere to put run-time debug
controls which may be inserted programmatically if needed.  To reflect that
it is not currently in use I have left it hidden in the
dlgProfilePreferences constructor.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>