BugFix: correctly destroy dockable map widgets on profile close
It has been a long-standing fatal issue:
https://github.com/Mudlet/Mudlet/issues/550
that dock-able mapper widgets were left orphaned with dangerously dangling
pointers that will crash Mudlet when a profile using them was closed whilst
multi-playing.  This commit makes the Host class now maintain a QPointer
that tracks the QDockWidget that this type of map has, and, as part of the
Host destructor ensures that the widget is given suicide instructions when
the Host itself is killed.

A QPointer was done so that should the map widget somehow get deleted by
other future code changes the pointer will automagically null itself so
that a test for it being non-null in the Host destructor is sufficient to
ensure that there is a widget that needs to be killed off at that time.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>