revise: use new Host::writeProfileData(...) method
Something similar to this method was going to be introduced in a different
PR by me and I saw the advantage of revising that to make it suitable for
this PR as well.

Have revised the original code to allow for a default 23 port number to
also be optional (and to reject the IANA reserved 0 port number) and to
include the currently needed null Host pointer check and revised the
run-time error message structure.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>