BugFix: create missing area when there is only one missing
A code structure error in a prior fix meant that instead of sorting the
missing areas if there was more than one such area BEFORE creating them
ALL, the corner case of a single missing area was not being created as
required.

Also move two separate QString::arg() usages into a single instance as
per new coding guidelines - as the first is a number it is necessary to
force it to a QString to avoid a compile failure (in later Qt versions ?)
that have a twin argument overload that expects a first number argument to
be followed by a second numeric value being a width specifier.

Also re-layout a few related console messages to avoid hitting a default
re-wrapping width at 80 characters.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>