Fix some TToolBar bugs. (#1026)
This PR focuses on making Floating Toolbars more usable and addressing bugs with the `TToolBar` object and its related parts.   In short this PR has made the following changes:  

1.  Fixes bug where Floating Toolbars were not closed with their Host Profile.
2.  Changes how Floating Toolbars are constructed, to avoid pulling the Editor Dialog away from the user abruptly.
3.  When Floating Toolbars are docked they get rebuilt in the same place you left them last, unless you delete them or restart the host profile.

Before this set of changes Floating Toolbars would create frustrating usability conditions.  
While a Floating Toolbar is active, clicking any of the Toolbars, Menus, or Buttons (`TAction`) items in the Editor Dialog window would result in the main window being raised and shown.  Thus any save, activation, or navigation event that resulted in a save of the items in the editor would result in the Editor Dialog effectively being pulled away from the user in the middle of them using it.
The larger portion of this PR focuses on that usability issue and makes changes to enforce building the Floating Toolbars only when the user has changed them in some way.