Multiple IRC Clients, one per host.
This changes the way IRC windows are managed by mapping IRC to a host
profile.  If no profile is loaded the IRC Client does not open.
When a Client gets messages it only posts those messages to its profile
rather than to all profiles.
IRC Client windows are closed when their profile is closed.