Enhance: make editor search have optional case-sensitivity and clear
Following on from https://github.com/Mudlet/Mudlet/pull/1096 I realised
that the search capability was still not case sensitive and given that Lua
is a case preserving language this struck me as a nuisance.  It is now
possible to configure this feature when searching with a button that
displays three line of letters "ABCD"/"AbCd"/"abcd" with all three lines
highlighted when searches are not case sensitive (i.e. a search for "A"
will also find "a") and only the middle line when the case does matter
(i.e. a search for "A" will not find the lower case one).

As we now have "boarder-outlines" around all the instances found in an
element in a lua script component, we ought to have a means to removed them
afterwards without having to click on another item and then return to the
current one being considered.  This commit also adds a clear button
to both clear the treeWidget_searchResults but also to remove those border
outlines.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>