BugFix: stop items being hidden in editor and then getting lost (Fix #1173)
I think that that absence of an initialisation value for the (bool)
mTemporary flag recent moved out of the Mudlet "items" derived classes and
down to the base class was causing these issues as "temporary" items are
not shown in the editor and not saved at the end of a session.  Technically
a value of not zero is a "true" value according to C coding standards but
C++ tends to use the specific integer values of 0 and 1 (not -1) for false
and true - I suspect there was some ambiguity when another value was being
seen when a derived item was instantiated without this boolean being set
to a definite value.

I have taken the liberty to also go through and mark the template file up
for `const`ness... and also put in C++14 `explicit` on the constructors
so that there is no automatic use of the one without the argument when a
"parent" argument is given and vice-versa.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>