Tweak: disable some items on key/value comboBoxs in Variable part of editor
Uses a means documented in https://stackoverflow.com/a/21376774/4805858
to replace the "standard" data model for the QComboBox with one from the
QListWidget which allows individual items to be flagged as disabled - which
the standard QComboBox does not.

This is designed to, in conjunction with a refactoring of code in
dlgTriggerEditor::slot_var_selected(...) to prevent invalid choices to be
made by the user (so types that they cannot enter through the Editor widget
but which are included in the type QComboBox s for display purposes:
functions for both and tables for one are disabled).

I would like to hide the editor widget when functions and tables are
selected in the tree widget - but that has proven to mess up the spacing
and layout for the remaining visible widgets and will have to be left until
I can bring: PR 436 "(release 30)bug fix get splitter working properly on
editor right side" into a usable state:
https://github.com/Mudlet/Mudlet/pull/436

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>