Revise: search option(s) now a menu with space for additional options
The mechanism now behaves very like the Qt Creator's own search options
which are presented as a clickable menu at the start of the search term
QLineEdit part of the editable QComboBox.

At present only Case Sensitivity is supported but it will be possible to
add additional search options - the most likely being whole word and
regular expression type searches. As the QtCreator inspiration does
different icons are shown depending on the options enabled on the menu
that is displayed is clicked (and encoded into the new enum/QFlags
"SearchOptions" which is stored in dlgTriggerEditor::mSearchOptions). As
new options are coded extra icons that support all the combinations will
be needed (although if a combination is not provided with an icon) the
grey binoculars icon that is used when no option is set is replaced with
a red one to indicate something is enabled and modifying the search
behaviour.

At present the options chosen are NOT stored against the search terms that
get stored into the search term QComboBox - it might be wanted to store
the search options chosen as part of each search term (use a item::data for
that) so they can be restore if an existing term is reused...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>