Revise: insert some missing casts to handle ambiguous un/signed char types
The "char" type is neither explicitly a signed or an unsigned type which
means that bitwise operations on each byte can go wrong unless it is
explicitly forced to one or the other.  In some cases modern compilers will
spot and warn about this and a couple of cases were detected in prior code.
This commit provides the missing forcing cast (to the Qt quint8 type) which
can be used for the bit-operations needed to process UTF-8 encoded bytes
required to remove the warnings and ambiguities that the (Clang) compiler
on Travis CI was moaning about.

I have also inserted codes to prevent clang-format from relaying-out the
codes where I have to use long, multi-part condition expressions as the
layout has been manually tweaked to try and group the terms in a more
readable structure which clang-format will try to undo if it is allowed to!

Also inserted some clang specific "things" that confirm that some
switch/case construct fall-through situations are wanted and not coding
errors to silence warnings that they otherwise produce in that compiler.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>