BugFix: correct a copy-paste then re-edit error in dlgTriggerEditor
Was producing a: "warning: comparison of constant -1 with boolean
expression is always true [-Wtautological-constant-out-of-range-compare]"
issue on Travis CI - which is just as well because the code in that one
case had not previous been correct. I think I must have been doing a
repetitive edit to code that I had copy and pasted and then revised and
messed up one instance.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>