BugFix: installModule(...) seg. fault after prior module manager usage (#1230)
* BugFix: installModule(...) seq. fault after prior module manager usage

This is a fix for the above Bug reported as:
https://github.com/Mudlet/Mudlet/issues/1224
there can now only be one module manager instance open at a time and a
QPointer<T> is maintained to it - however the QTableWidget that displayed
the module data was pointed to by something not implemented as a smart
pointer so it was dangling when the dialog was closed and that pointer
was used by (bool) mudlet::moduleTableVisible() which was called by the
lua function installModule(...) and thus provoked a segment fault.

The fix here is multi-pronged and addresses a slightly wider scope in that
I have made all the mudlet class member pointers to elements in this the
module manager AND the unrelated but similarly designed package manager
to be smart QPointer<T>s so they all become nullptrs when the corresponding
QDialog ceases to exist - and if there are any improper uses of those
pointers again it should be immediately apparent that a nullptr has been
used instead of a dangling one.

In reviewing the module manager it also became apparent that it was also
unsafe in that the individual dialog slot functions to add/remove/get help
on modules all used the currently active profile to add/remove/get help
for the modules on - which is dodgy because if multi-playing that
assumption is wrong.  To fix this when the module manager dialog is created
it also records the profile that was active at the time and the remaining
methods only use that after checking it (as a QPointer<Host> type) is
still valid.  The method that records the profile:
(void) mudlet::slot_module_manager()
now also arranges for the "destroyed" SIGNAL from the dialog to call a new
SLOT (void) mudlet::slot_module_manager_destroyed() which ensures the
pointer that recorded the profile that initiated the dialog is correctly
cleared so another profile can use the dialog.

It seemed to me to actually be useful to include the profile concerned in
in the module manager dialog in the title-bar.

Whilst I was rooting around in the module manager I also fixed one or two
other things:
* https://github.com/Mudlet/Mudlet/issues/1187 "Module manager says "sync?"
  when first opened" In fact there is no need for any text there at all the
  checkbox is sufficient to show the state of the setting and by shortening
  the Heading for that column from "Save & Sync" to just "Sync" it was
  possible to narrow that column a little so that the checkboxs in that
  column do not seem lonely without any text to the right of them.
* I was able to QStringLiteral/QLatin1String wrap some raw string literals
  that are not GUI text QString in the touched methods and some new nearby
  gamepad code.
* Revised the tool-tip for the "Sync" checkboxes in the module manager
  dialog.
* I was able to redesign the dialog layout to be a little simpler and make
  better sub-division of the space available. I also manage to clean up
  the explanation text in the bottom of the dialog so it all shows by
  and also to clean up the HTML to be simpler - however as of Qt Creator
  4.3.1 the form designer either as a plugin or as a stand-alone utility
  now completely fails to retain the "simple" HTML option on the text
  design/entry widget it uses so as soon as a different element on the
  form is selected the HTML is immediately reverted to the horrendously
  complex HTML it likes to use but which due to Qt still not having
  addressed https://bugreports.qt.io/browse/QTBUG-1136 makes it almost
  impossible to translate with Qt Linguist...

The default size set was only 7 and 9 for the "help" text - this thus
appears smaller that nearly everything else so the addition step was made
to remove explicit font size setting in the module_manager dialog...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>