Fixup: improve some things reported by Codacy
Added a hash-bang entry to src/mudlet-lua/genDoc.sh

Commented out "almost illegal" #define of _DEBUG_ to nothing in:
src/dlgConnectionProfiles.cpp as that is reserved for compilers' use in the
C++ standards and should not be #define-d by the end user.

Removed some lines containing just a '#' character in
3rdparty/communi/include/IrcCore/ircglobal.h as Codacy was reporting an
un-closed #if...#else - the "#  endif" was present but it is not clear that
isolated '#' characters are wise/safe to use in a C/C++ source code file.

Also added some non-source code files to main Qt project files including
one modified by this commit so that they show up in the Qt Creator IDE.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>