BugFix: #1267 deleted item crash if profile preferences Editor previews it
Whilst reviewing Coverity items CID1449137/38/39/41 I found that it was
possible to crash Mudlet if an item is deleted in the main editor after
the profile preferences is opened and then that item is selected for
the editor preview on the "Editor" tab of that dialog.

This should address the issue.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>