Allow to register lua function objects to events
Before it was only possible to register lua code as strings to events.
This resulted in either being unable to use closures for events (by writing
code as anonymous event handler) or littering the global namespace (by
giving a global function name as event handler).

This solution trades a single global function (as event handler) for the
ability to use local functions (or even lambdas) as event handlers.