Refactor: remove unneeded dialog titles from some forms + other tweaks
Several forms/dialogues do not need a title (because they are incorporated
into other forms; others that are floated free have their title set
elsewhere - in both cases it is advantageous to remove strings that do NOT
need to be made translatable when Mudlet is localised into a language and
they just pollute the code/make extra strings that would automatically be
marked as needing translation - so removing them is mildly useful.

The "composer" form/dialog used for certain MUDs had the title:
"News & Message Composer" whereas it is simpler for translation to turn the
"&" back to "and" as that character has to be escaped in XML files such as
the ".ts" ones we will use to make "translations".

The default view in the "about_dialog" was defaulting to showing the third
tab (the "3rd Party Code Licences" one) compared to the intended first one
("About Mudlet") - this commit restores it to the latter.

A commented out line in dlgSourceEditorArea.cpp can also be removed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>