Handle string event handlers in lua as well
To allow the removal (killing) of string event handlers, they are now also
handled in lua. This leads to the C++ implementation of
registerAnonymousEventHandler() being almost unused. Now it should be used by
the overwriting lua code only.

This also adds a last optional parameter oneShot to the
registerAnonymousEventHandler() function. With that option set, the event
handler will be automatically removed after one call.