Refactor: reduce use of raw string literals for directories and file paths
Using a static method means that we can centralise the generation of
strings used to access parts of the file system used to hold Mudlet user
data - which may be useful if we change them for - say running from a
thumb-drive or just to use an OS specific location for them. {Using a
".config/mudlet" sub-directory on Windows is not quite how that OS expects
things to be done...!}

This also reduces the number of raw strings that will have to be wrapped
in QStringLiteral(...) and the like for forthcoming I18n work.

Also:
* The use of "mudlet-data" as a windows symlink sub-directory to
  ".config/mudlet" in the user home directory may not work because the Qt
  Documentation suggests that a ".lnk" extension is required - this commit
  adds that FOR THAT OS ONLY.

* Previous code placed (only some - for an unexplained reason not all the
  files were being copied) the DejaVu fonts that we include in the
  resources directly into the main mudlet user data folder - but the
  included documentation including the COPYRIGHT.TXT and other files ONLY
  relate to those font files and NOT to Mudlet as a whole.  I have taken
  the liberty of moving and loading them from a ./fonts/ subdirectory which
  makes more sense.  I have noted that, at least, the Debian Linux
  packager actually removes the font files from their configuration because
  they already include a shared system wide copy of the same fonts and in
  a much newer (2.37 at present) version compared to the 1.10 that we
  bundle - perhaps our code needs to check for the existence/availability
  of the fonts before we try to load ours into the Mudlet application at
  run-time!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>