Enhance: enable building on FreeBSD
Some of the LIBS and INCLUDEPATH setting in the QMake project file need to
be tweaked to allow building on the FreeBSD OS - fortunately there is a
test-variable "freebsd" that is true when qmake parses its project files
which can be used to detect and allow for that platform.

It is my intention of "porting" Mudlet to FreeBSD and, as Mudlet is a GPLed
project it can not be included in the BSD licenced "core" of that OS but
should be submitted to https://www.freebsd.org/ports/ for inclusion the
"/usr/ports" hierarchy of the project (should this be successful then
"/usr/ports/games/mudlet/" on any FreeBSD system will contain a platform
specific Makefile and some other files which means that a system
administrator {"root" user} can go to that sub-directory and run the
command "make install" and the system will download the specified mudlet
tarball, unpack it and after building any other pre-requisites will build
and install the mudlet executable as "/usr/local/bin/mudlet" and install
other required files in suitable locations in the "/usr/local" area of the
file-system).

At present I have not managed to build a working executable with GCC
versions 5.4.0 and 6.3.0 but I am building 7.2.0 at the moment...

In examining the qmake project file it became apparent that some previous
tweaks to the non-mcvc compliers (GCC or Clang) were only being applied to
the C++ sources but we do use a few C files as well so I modified those
option changes to apply to both types of source files.

Checking the FreeBSD file hierarchy it became clear that it too
supports/uses the FreeDeskTop directory system so in the absence on Unix
systems that are not MacOS it is not unreasonable to use the highest
(first) specified directory in the XDG_DATA_DIRS environment variable as
a value for the PREFIX environment variable which is the base directory
for where Mudlet should place particular files on *nix OSes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>