Revise: fixup past windows symlink error and other suggested revisions
This commit takes steps to actually repair the sym-link error that the
previous commit merely avoids creating.  When it does so it also pops-up
a warning but as this should only ever happen once in an installation and
only if a previously bugged version has been run by the user concerned I
think this is acceptable - especially if it doesn't work as expected on
an OS platform especially that I cannot currently test myself!

The new function in main to install the font-files if they do not exist
has been revised from "checkAndCopyFont" to a shorter "copyFont" though
it is otherwise unchanged.

A comment in (int) TLuaInterpreter::getMudletHomeDir(lua_State*) has been
removed - for the record it indirectly referred to an issue mentioned in
the Forums at:
https://forums.mudlet.org/help-forum-f9/getmudlethomedir-and-encoding-t4508.html
which this should finally (several years later) fix...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>