Allow to register lua function objects to events (#1284)
Before it was only possible to register lua code as strings to events.
This resulted in either being unable to use closures for events (by writing
code as anonymous event handler) or littering the global namespace (by
giving a global function name as event handler).

This solution trades a single global function (as event handler) for the
ability to use local functions (or even lambdas) as event handlers.
This also adds a last optional parameter oneShot to the
registerAnonymousEventHandler() function. With that option set, the event
handler will be automatically removed after one call.