Fixed a rare crash with search
In the case someone repeatedly searched for the same thing, the last search result happened to be right at the top of the buffer, and then the buffer was shrunk because some lines were batch deleted, the index to start searching lines from was never adjusted - and caused a crash.

This should fix it by choosing a minimum available value of the total lines or the last index to search from.