BugFix: module/package handling error
In #1286 I made what I think is a copy paste error and ended up with
the wrong string being used in:
(bool) Host::installPackage(const QString&, int)
for the directory to be used to unzip (archive) file format
packages/modules (there was a '/' missing between the profile's HOME
directory name and the "packageName" of the concerned module being the
file name contained in the "config.lua" should that differ from the base
name of the archive file).

For archives where all the files are contained in a single directory the
effect is to create a folder/file in the wrong place in the mudlet part of
file-system and to expand all the modules in the profile's HOME directory
(the fact that they all contain a file called "config.lua" is not an issue
as the modules are unpacked in sequence and that file is read for each file
before being overwritten by the next package) what will be problematic is
a newly installed package will not have been expanded and will not be in
the right place if a bugged version of mudlet is the only one that has
been used.

For modules that contain sub-directories this bug makes an even bigger mess
in the profile's home directory... 8-(

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>