revise: move to using Qt::PenStyle instead of QString for custom lines
Storing a QString as the line style for any exit with a custom line does
not work so well when the language used to display the option has to be
changed to accommodate a different language.  This commit changes the
TRoom class to store the Qt:PenStyle enum value directly instead of a
representation of the name of the style.

This did show up a slightly tricky issue when serialising the enum as at
some point in the Qt 5.x history it seems to have gotten harder to access
the underlying integer value.  This is solved by a pair of QDataStream
handling "friends":
* (QDataStream) &operator<<(QDataStream&, const Qt::PenStyle&) {writing}
* (QDataStream) &operator>>(QDataStream& ds, Qt::PenStyle& value) {reading}
they have been included at the top of the TMap and TRoom classes
respectively and which are used "behind" the scenes when a Qt::PenStyle is
serialised.  NOTE: the absence of the function in TRoom will throw up a
slew of compilation errors about "no match for 'operator>>' (operand types
are 'QDataStream' and 'QMap<QString, Qt::PenStyle>::mapped_type {aka
Qt::PenStyle}') in "qdatastream.h" as line 284 but it will be unclear
otherwise what on earth has gone wrong.  More insidiously the absence of
a matching function in the TMap write to file process will silently
seem to work but write out a different number of bytes (several rather than
one or two) such that any data afterwards in the file produced will be
corrupted...!!!

When I tried to work out what texts to use for other languages for the
different line styles I realised I could not be certain of the Google
translations so I ran up a set of icons to use as well - they are a bit
skimpy at the moment so may need a bit of tweaking to look better.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>