BugFix: correct broken WIN32 #defines and use of libzip 1.x functions
Repairs a stupid mistake in #ifdef-ing a couple of Windows specific chunks
of code I made before.

I didn't realise it but I am using a (locally compiled) version 1.x of
libzip but some Debian versions and the Trusty Ubuntu Travis Platform only
comes with an 0.1x version by default (that may not encode UTF-8 paths)
correctly - I am not sure of the exact details) although the source for
libzip {https://nih.at/libzip/libzip.html} has declared the functions that
we were using {zip_add(...) and zip_add_dir(...)} OBSOLETE - however that
is all that is available for those cases so I have added #ifdefs that will
use the later functions {zip_file_add(...) and zip_dir_add(...)} if they
are available.  Whilst this hits two places in the previous part of the PR
there is one more instance in the code base that I have also fixed up in
this way.

As it happens both of these issues have arisen as a result of a reworking
of the package/module handling code which, prior to this PR will silently
FAIL to make packages/modules if there are any files placed in the staging
area in a sub-directory.  That staging area was described as a "temporary"
one but in fact persists beyond the production of a package/module. As part
of the reworking it has been made to have the option (NOT ENABLED BY
DEFAULT) to clean up the temporary area afterwards but if a package/module
is being developed then it is likely that the other resources are likely
wanted to be kept for producing later, updated/revised, ones...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>